backend/requirements.txt
anthropic
fastapi
uvicorn[standard]
python-dotenv
websockets
httpx
mcp
